search for feature files
