+++ lib/gherlint/GherLint.js
@@ -1,18 +1,21 @@
 const fs = require("fs");
-const path = require("path");
+const glob = require("glob");
+const { extname } = require("path");
 
 const { Linter } = require("../linter");
 const { gherlintrc: defaultConfig } = require("../config");
-const { isDir } = require("../../utils/Path");
+const Path = require("../../utils/Path");
 
 module.exports = class GherLint {
     static {
         GherLint.linter = new Linter(defaultConfig);
         GherLint.problems = {};
+        GherLint.FEATURE_FILE_EXTENSION = ".feature";
     }
 
     static async lintFiles(paths) {
-        GherLint.resolvePaths(paths).forEach((file) => {
+        GherLint.getFeatureFiles(paths).forEach((file) => {
+            console.log(file);
             const text = fs.readFileSync(file, "utf8");
 
             const problems = GherLint.linter.lint(text);
@@ -27,19 +30,24 @@ module.exports = class GherLint {
         GherLint.problems[path] = problems;
     }
 
-    static resolvePath(file) {
-        const realPath = path.resolve(process.cwd(), file);
-        if (isDir(realPath)) {
-            throw new Error(
-                "Please, provide a single feature file or a glob pattern. We don't support directories yet."
-            );
-        }
-        return realPath;
+    static getFeatureFiles(paths) {
+        const features = [];
+        paths.forEach((path) => {
+            features.push(...GherLint.searchFeatureFiles(path));
+        });
+
+        return features;
     }
 
-    static resolvePaths(paths) {
-        return paths.map((file) => GherLint.resolvePath(file));
+    static searchFeatureFiles(pattern, files = []) {
+        glob.sync(pattern, { absolute: true, cwd: Path.cwd() }).map((match) => {
+            if (Path.isDir(match)) {
+                return GherLint.searchFeatureFiles(`${match}/*`, files);
+            }
+            if (extname(match) === GherLint.FEATURE_FILE_EXTENSION) {
+                return files.push(match);
+            }
+        });
+        return files;
     }
-
-    static readConfigFile() {}
 };
